♻️ refactor the auth module
