+++ apps/web/components/page-components/auth/SignIn.tsx
@@ -1,8 +1,7 @@
 "use client";
 
-import { FormControl, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
 import Link from "next/link";
+import MyFormField from "../../base/MyFormField"
 
 interface FormType {
   [key: string]: any;
@@ -26,40 +25,8 @@ function renderSignIn({
   if (isPhoneMode) {
     return (
       <div>
-        <FormField
-          control={form.control}
-          name="phone"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>手机号码</FormLabel>
-              <FormControl>
-                <Input
-                  className="input-bordered"
-                  {...field}
-                  placeholder="请输入手机号码"
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="code"
-          render={({ field }) => (
-            <FormItem className="form-control mt-[5px]">
-              <FormLabel className="label">验证码</FormLabel>
-              <FormControl>
-                <Input
-                  className="input-bordered"
-                  type="code"
-                  placeholder="请输入验证码"
-                  {...field}
-                />
-              </FormControl>
-            </FormItem>
-          )}
-        />
+        <MyFormField form={form} name={"phone"} placeholder={"请输入手机号码"} label={"手机号码"}/>
+        <MyFormField form={form} name={"code"} placeholder={"请输入验证码"} label={"验证码"}/>
         <button
           className={`btn btn-outline btn-error mt-2`}
           onClick={handleClick}
@@ -83,40 +50,9 @@ function renderSignIn({
   } else {
     return (
       <div>
-        <FormField
-          control={form.control}
-          name="username"
-          render={({ field }) => (
-            <FormItem>
-              <FormLabel>用户名/邮箱</FormLabel>
-              <FormControl>
-                <Input
-                  className="input-bordered"
-                  {...field}
-                  placeholder="请输入用户名/邮箱"
-                />
-              </FormControl>
-              <FormMessage />
-            </FormItem>
-          )}
-        />
-        <FormField
-          control={form.control}
-          name="password"
-          render={({ field }) => (
-            <FormItem className="form-control mt-[5px]">
-              <FormLabel className="label">密码</FormLabel>
-              <FormControl>
-                <Input
-                  className="input-bordered"
-                  type="password"
-                  placeholder="请输入密码"
-                  {...field}
-                />
-              </FormControl>
-            </FormItem>
-          )}
-        />
+        <MyFormField form={form} name={"username"} placeholder={"请输入用户名/邮箱"} label={"用户名/邮箱"}/>
+        <MyFormField form={form} name={"password"} placeholder={"请输入密码"} label={"密码"}/>
+         
         <label className="label justify-end">
           <Link
             href="#"
